requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
Flask>=2.0.0
tqdm>=4.62.0
tensorboard>=2.12.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.4.0
kagglehub>=0.2.0
werkzeug>=2.0.0 